Switch skull stats to match new repair stats
Skull just counts towards repair now, not base stats
Armor is now deprecated for removal, and is hidden from the tooltip unless non-zero. All base skulls had their armor value removed

+++ src/main/java/slimeknights/tconstruct/tools/stats/RepairStats.java
@@ -28,7 +28,7 @@ public record RepairStats(MaterialStatType<?> getType, int durability) implement
 
   @Override
   public List<Component> getLocalizedInfo(float scale) {
-    return List.of(IToolStat.formatNumber(REPAIR_KEY, ToolStats.DURABILITY.getColor(), this.durability * scale / MaterialRecipe.INGOTS_PER_REPAIR));
+    return List.of(IToolStat.formatNumber(REPAIR_KEY, ToolStats.DURABILITY.getColor(), (int)(this.durability * scale / MaterialRecipe.INGOTS_PER_REPAIR)));
   }
 
   @Override
